Move tiled.database -> tiled.auth_database.
